Planning: OpenSpace: refractor folders structure
