+++ modules/planning/open_space/coarse_trajectory_generator/hybrid_a_star.cc
@@ -18,7 +18,7 @@
  * @file
  */
 
-#include "modules/planning/open_space/hybrid_a_star.h"
+#include "modules/planning/open_space/coarse_trajectory_generator/hybrid_a_star.h"
 
 namespace apollo {
 namespace planning {
