+++ modules/planning/open_space/coarse_trajectory_generator/node3d.h
