+++ modules/planning/open_space/coarse_trajectory_generator/reeds_shepp_path.h
@@ -30,7 +30,7 @@
 #include "cyber/common/macros.h"
 #include "modules/common/configs/proto/vehicle_config.pb.h"
 #include "modules/common/math/math_utils.h"
-#include "modules/planning/open_space/node3d.h"
+#include "modules/planning/open_space/coarse_trajectory_generator/node3d.h"
 #include "modules/planning/proto/planner_open_space_config.pb.h"
 
 namespace apollo {
