+++ modules/planning/open_space/coarse_trajectory_generator/reeds_shepp_path_test.cc
@@ -18,7 +18,7 @@
  * @file
  */
 
-#include "modules/planning/open_space/reeds_shepp_path.h"
+#include "modules/planning/open_space/coarse_trajectory_generator/reeds_shepp_path.h"
 
 #include <memory>
 
@@ -27,7 +27,7 @@
 #include "modules/common/configs/vehicle_config_helper.h"
 #include "modules/common/util/util.h"
 #include "modules/planning/common/planning_gflags.h"
-#include "modules/planning/open_space/node3d.h"
+#include "modules/planning/open_space/coarse_trajectory_generator/node3d.h"
 #include "modules/planning/proto/planner_open_space_config.pb.h"
 
 namespace apollo {
