+++ modules/planning/open_space/constraints_formulation/BUILD
@@ -1,0 +1,28 @@
+load("//tools:cpplint.bzl", "cpplint")
+
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "open_space_ROI",
+    srcs = [
+        "open_space_ROI.cc",
+    ],
+    hdrs = [
+        "open_space_ROI.h",
+    ],
+    copts = ["-DMODULE_NAME=\\\"planning\\\""],
+    deps = [
+        "//cyber/common:log",
+        "//modules/common/configs:vehicle_config_helper",
+        "//modules/common/vehicle_state/proto:vehicle_state_proto",
+        "//modules/map/hdmap:hdmap_util",
+        "//modules/map/pnc_map",
+        "//modules/planning/common:frame",
+        "//modules/planning/common:indexed_queue",
+        "//modules/planning/common:obstacle",
+        "//modules/planning/common:planning_gflags",
+        "@eigen",
+    ],
+)
+
+cpplint()
