+++ modules/planning/open_space/open_space_trajectory_generator.h
@@ -31,9 +31,9 @@
 #include "modules/common/math/vec2d.h"
 #include "modules/common/vehicle_state/proto/vehicle_state.pb.h"
 #include "modules/planning/common/frame.h"
-#include "modules/planning/open_space/distance_approach_problem.h"
-#include "modules/planning/open_space/dual_variable_warm_start_problem.h"
-#include "modules/planning/open_space/hybrid_a_star.h"
+#include "modules/planning/open_space/trajectory_smoother/distance_approach_problem.h"
+#include "modules/planning/open_space/trajectory_smoother/dual_variable_warm_start_problem.h"
+#include "modules/planning/open_space/coarse_trajectory_generator/hybrid_a_star.h"
 #include "modules/planning/proto/planner_open_space_config.pb.h"
 #include "modules/planning/proto/planning_config.pb.h"
 #include "modules/planning/proto/planning_internal.pb.h"
