+++ modules/planning/open_space/trajectory_smoother/distance_approach_ipopt_interface.cc
@@ -18,7 +18,7 @@
  * @file
  */
 
-#include "modules/planning/open_space/distance_approach_ipopt_interface.h"
+#include "modules/planning/open_space/trajectory_smoother/distance_approach_ipopt_interface.h"
 
 namespace apollo {
 namespace planning {
