+++ modules/planning/open_space/trajectory_smoother/distance_approach_ipopt_interface.h
