+++ modules/planning/open_space/trajectory_smoother/dual_variable_warm_start_problem.cc
@@ -18,7 +18,7 @@
  * @file
  */
 
-#include "modules/planning/open_space/dual_variable_warm_start_problem.h"
+#include "modules/planning/open_space/trajectory_smoother/dual_variable_warm_start_problem.h"
 
 #include <algorithm>
 #include <iomanip>
