+++ modules/planning/open_space/trajectory_smoother/dual_variable_warm_start_problem.h
@@ -1,5 +1,5 @@
 /******************************************************************************
- * Copyright 2017 The Apollo Authors. All Rights Reserved.
+ * Copyright 2018 The Apollo Authors. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,43 +20,29 @@
 
 #pragma once
 
-#include <algorithm>
-#include <iomanip>
-#include <utility>
 #include <vector>
 
 #include "Eigen/Dense"
-#include "IpIpoptApplication.hpp"
-#include "IpSolveStatistics.hpp"
 
-#include "modules/common/time/time.h"
-#include "modules/planning/common/planning_gflags.h"
-#include "modules/planning/open_space/distance_approach_ipopt_interface.h"
+#include "modules/planning/open_space/trajectory_smoother/dual_variable_warm_start_ipopt_interface.h"
 #include "modules/planning/proto/planning.pb.h"
 
 namespace apollo {
 namespace planning {
 
-class DistanceApproachProblem {
+class DualVariableWarmStartProblem {
  public:
-  explicit DistanceApproachProblem(
-
+  explicit DualVariableWarmStartProblem(
       const PlannerOpenSpaceConfig& planner_open_space_config);
 
-  virtual ~DistanceApproachProblem() = default;
+  virtual ~DualVariableWarmStartProblem() = default;
 
-  bool Solve(const Eigen::MatrixXd& x0, const Eigen::MatrixXd& xF,
-             const Eigen::MatrixXd& last_time_u, const size_t& horizon,
-             const double& ts, const Eigen::MatrixXd& ego,
-             const Eigen::MatrixXd& xWS, const Eigen::MatrixXd& uWS,
-             const Eigen::MatrixXd& l_warm_up, const Eigen::MatrixXd& n_warm_up,
-             const std::vector<double>& XYbounds,
-             const size_t& obstacles_num,
+  bool Solve(const size_t& horizon, const double& ts,
+             const Eigen::MatrixXd& ego, const size_t obstacles_num,
              const Eigen::MatrixXi& obstacles_edges_num,
              const Eigen::MatrixXd& obstacles_A,
-             const Eigen::MatrixXd& obstacles_b, Eigen::MatrixXd* state_result,
-             Eigen::MatrixXd* control_result, Eigen::MatrixXd* time_result,
-             Eigen::MatrixXd* dual_l_result, Eigen::MatrixXd* dual_n_result);
+             const Eigen::MatrixXd& obstacles_b, const Eigen::MatrixXd& xWS,
+             Eigen::MatrixXd* l_warm_up, Eigen::MatrixXd* n_warm_up);
 
  private:
   PlannerOpenSpaceConfig planner_open_space_config_;
