+++ modules/tools/open_space_visualization/distance_approach_problem_wrapper.cc
@@ -18,9 +18,9 @@
  * @file
  **/
 
-#include "modules/planning/open_space/distance_approach_problem.h"
-#include "modules/planning/open_space/dual_variable_warm_start_problem.h"
-#include "modules/planning/open_space/hybrid_a_star.h"
+#include "modules/planning/open_space/trajectory_smoother/distance_approach_problem.h"
+#include "modules/planning/open_space/trajectory_smoother/dual_variable_warm_start_problem.h"
+#include "modules/planning/open_space/coarse_trajectory_generator/hybrid_a_star.h"
 
 namespace apollo {
 namespace planning {
